Disable rubocop analysis on codeclimate
We do it during the build so it adds little to do it
again here. This decision was influenced by the fact that
we can't easily pull in the before_commit .rubocop.yml
for codeclimate to use (the dieasl solution being to use
inherit_from <raw beforecommit rubocop.yml url>
from a .codeclimate_rubocop.yml
but this doesn't work because network connections are
unavailable on hosted codeclimate.
Also remove the rubymotion engine as that was a stray.