Replace before_commit; update Rails and ReadMe
Key changes:
- Replace before_commit with individual CI checks;
- Update Rails to 4.2.7.1;
- Expand and update the ReadMe for this project.

[before_commit](https://github.com/EnvironmentAgency/before_commit) was an attempt to have a central gem which could be included in every Ruby based project. The gem would manage the checks we wished to perform in each, thereby ensuring consistency and quality across the projects.

In practise we have found when it works, it works great however when it doesn't, it can be very confusing to try and resolve what is actually failing (the project, the check, or **before_commit** itself). We also frequently encountered issues with trying to essentially enforce a centrally controlled workflow on differing development environments.

We have agreed instead to move to ensuring the checks are run independently, and only on the relevant CI server. It will be the responsibility of devs to run these same checks locally, in the way appropriate to their environment, to ensure they don't '**break the build**'.

This change covers removing the dependency on **before_commit** and instead moving to the new agreed model. This involved

- removing the .git-hooks and [Overcommit](https://github.com/brigade/overcommit) (the gem **before_commit** is built on) config and gemfiles from the project
- removing them as dependencies in the project
- removing the require statement and loading of Rake tasks from the `Rakefile`

TODO:
- re-visit i18n-tasks and its reporting of missing keys.