Add design & accessibility improv. for OS Places
Having presented the initial version of adding the OS places notice and T&C's to the service, this was then evaluated by the team's content designer and quality analyst.

To improve accessibility it was recommended that

- 'Terms and Conditions' in the `os_places_terms.html.erb` page was switched from `<p>` to a `<h2>`.

To improve the design of the notice it was recommended that

- it was moved below the submit button and made visibly distinct from the main content
- the text size was reduced
- the link to OS places was removed
- 'OS' was expanded to 'Ordnance Survey'

To improve the design of the terms page it was recommended that

- Initial references to Ordinance Survey and Open Government License are postfixed with their acronym (OS) and (OGL)
- The link to OS places is removed
- 'Terms and Conditions' casing is corrected (lowercase the 'c' in conditions)

Finally where we still have links the `rel` attribute was removed along with the use of '(opens in a window)' in the displayed text. GDS have deprecated this behaviour so it is no longer needed here. You can read more at https://github.com/alphagov/govuk_frontend_toolkit/pull/293.