Remove calls to html_safe / add opens new window
As per last pieces of feedback, removed the explicit calls to `.html_safe` in the _postcode.html.erb and os_places_terms.html.erb pages, and instead added a `_html` postfix to the relevant keys in the localisation files (see http://guides.rubyonrails.org/i18n.html#using-safe-html-translations for more details).

Also added back in the text (opens new window) to the link in the notice in order to keep it consistent with the rest of the site (but the `rel` link remains removed).