Update codeclimate api key in .travis.yml
The connection between codeclimate and the repo in github was broken and the only solution appears to be to delete the project in codeclimate, and then re-add it.

This has caused an update to the api key, which is used by Travis-Ci to allow it to push test coverage results to codeclimate. This change simply updates the key in the `.travis.yml` file.