Fix broken email test
Following the bundle update `enrollment_mailer_submitted_spec.rb` broke, specifically `EnrollmentMailer submitted email content in html format pulls in translated content for .foo`.

Still haven't determined what, but essentially the email content is escaping content like "We'll contact you..." to "We&#39;ll contact you...", but not content like "Telephone: <a href="tel:03708506506">03708 506 506</a>".

So the matcher was no longer able to find what `t(key)` was returning in some cases. Escaping all keys wasn't a solution because that broke those that featured html. So it was either a choice of splitting out into 2 separate tests for the 2 types of content, or going with the solution implemented here.