Update dependencies (#260)
We had a lot of out-of-date gems, which means security vulnerabilities. This PR will cover those updates.

I decided to lock Rubocop to the existing version as updating it would have raised a lot of linter issues. Long-term we should switch to using https://github.com/DEFRA/defra-ruby-style but that'll need a separate PR (or several).

* Update Rails
* Update ea-area_lookup - this includes several minor versions of Nokogiri with security fixes.
* Update bootstrap-sass - there's a security vulnerability in the old version.
* Remove mailcatcher from dependencies - this removes a security warning from one of its dependencies. We don't need it here anyway as we run it separately through the new Vagrant box config.

We then had to fix the build as there was a problem with requiring the Virtus dependency.

The tests were also failing on the new version of finite_machine, so locked it to the version we were previously using.