Add github changelog generator to the project
As a team our release process now relies on updating a Changelog, from which we can then extract the detail needed for raising our Requests for Change.

FRAE was put together before this new team process was put in place, so doesn't have the [GitHub Changelog Generator](https://github.com/github-changelog-generator/github-changelog-generator) gem as a dev dependency.

This change adds it, along with a rake task to generate the file, and the first version of CHANGELOG.md