Remove documentation regards the state machine
The team that are now maintaining this have a repo where they collate information on their ways of working and the services they are working on ([ruby-services-team](http://github.com/DEFRA/ruby-services-team)).

We want all documentation to also be collated there, and as such we have already migrated a [copy of this document](https://github.com/DEFRA/ruby-services-team/blob/master/services/frae/state_machine.md).

So to avoid issues with having the content duplicated, this change removes STATE_MACHINE_README.md from the repo.