Fix broken tests (#287)
When running the tests recently spotted that any that relied on the address lookup were failing. A review of the cassettes highlighted that they have not been rerecorded since May 2016, and it looked like it was something in the response the tests were getting that was causing the error.

So this makes 2 changes

- it re-records the responses to bring them up to date
- it deletes the responses that were hitting our external Azure instance of ABF

The Azure instances were used by the project till VCR was added. As the cassettes are locked and the requests are based on whatever is set in the ENV VARS, there seemed no point in maintaining 2.

Finally as this would mean updating the env vars in Travis, we took the opportunity to add them to `.travis.yml` rather than have them hidden away in the travis config online.