Update defra ruby area to handle multiple results (#295)
https://eaflood.atlassian.net/browse/RUBY-484

In preparation for testing the working area lookup integration our awesome QA @andrewhick was able to find examples of coordinates which when used would cause the Water Management WFS which we rely on to return multiple results for the admin area.

We thought initially we'd be fine, and we'd just get the data from the first result. As the previous implementation was not handling multiple results this would be ok. However when we checked the [defra-ruby-area gem](https://github.com/DEFRA/defra-ruby-area) we found this was not the case.

In these situations it would simply concatenate the data. So if the first result returned `North` and the second result returned `East` the result the FRAE engine would see would be `NorthEast`.

We decided to take this opportunity to update the the gem to properly handle multiple results, which means its API changed. So this change covers updating the FRAE engine to match this update in the gem's API.