Update water mgmt area job to report errors
Prior to this change if an error occurred when the engine attempted to lookup the water mangement area for an exemption nothing was logged.

This would mean we on the service team would no longer be aware of any issues with the WFS service until the users reported it.

This change adds reporting of errors to Errbit, and some tweaks to the job class to support it.