Do not generate postcode in address factory
We have found the `spec/lib/tasks/lookups_spec.rb` in the [flood-risk-back-office](https://github.com/DEFRA/flood-risk-back-office) is failing sometimes due to the postcode `Faker::Address.postcode` sometimes returns.

You can also see we've had issues before with invalid or problematic postcodes returned from [Faker](https://github.com/faker-ruby/faker) hence a _whole_ bunch of code to try and generate our own valid postcode.

This change set out to swap the Faker call with a call to `generate_postcode()` however that then caused a test in the engine to always fail (`spec/controllers/flood_risk_engine/enrollments/steps_controller/partnership_details_spec.rb:57` for reference).

😩😩😩

We know the factories and rspec are not set up as we would normally do in our other projects, and at some point we want to update it to be consistent. So it doesn't warrant a deep dive investigation. Also we would not want to change too much in the pursuit of fixing it.

The simple solution is actually just to always use a known valid UK postcode. In this case having a generated valid postcode is not telling us anything a static one cannot, so with this change

- we can always be sure we are getting a known valid postcode
- the logic is simpler
- we get to delete a whole bunch of code that had no tests itself