Add comment about no working address scope (#304)
https://eaflood.atlassian.net/browse/RUBY-604

Spotted whilst we were trying to deliver an automated job to update the Area on site locations that are missing one.

We added the scope to ensure we were only updating locations linked to site addresses, but when testing the job we spotted

- there is an existing bug in the service which means the address type is never set to anything other than `:primary`
- locations only get created for site addresses

To fix the area lookup just means reverting the introduction of the scope in the back office, but we decided against deleting the scope altogether. The code is correct and would work of the bug did not exist. So rather than delete something of value, we'll leave it whilst we decide what to do.