Move defra ruby alert config into engine
By including config in the engine to set the config in defra-ruby-alert, we can manage it in one place and reduce the duplication still further.

We've used this pattern in WEX and WCR and it means we handle the stuff that will be the same across apps, and just require them to pass the info that's unique to them.