Use defra-ruby-alert for airbrake mgmt (#312)
https://eaflood.atlassian.net/browse/RUBY-604

RUBY-604 is about adding a rake task which finds any exemptions with a missing water management area and updates them. The intention is to run this overnight.

When we added this we simply ported an implementation we had done that was very similar from Waste Exemptions and used a pattern we have since adopted for all our rake tasks.

Part of the pattern is to ensure Airbrake is closed before the rake task completes. It had been spotted on another service that because Airbrake is asynchronous if the app closes before the error gets reported, we will never hear about it. So if our rake task were to fail overnight, we wouldn't know.

However because FRAE is on an older version of Airbrake, what we didn't know when we completed PR #152 was that calling `Airbrake.close` would cause an error when run in production mode.

So in the case of FRAE we would need to wrap it in a begin/rescue, and this suddenly becomes a reasonable amount of code we could be copying and pasting everywhere. Plus we'd done some additional thinking about the problem calling `Airbrake.close` was solving. Yes, if an error occurs during a rake task we want to know about it. But this also applies to the main rails app. What if we encountered an error that caused Rails to close suddenly? There would be a chance the error would not be reported because Airbrake would not have had a chance to report it.

So after some experimentation, we found the solution to both problems was to call `Airbrake.close` in an `at_exit` hook. And because this is a pattern we'd want to apply to all our services, we created [defra_ruby_alert](https://github.com/DEFRA/defra-ruby-alert) to manage both the hook and configuring airbrake.

So long story short, this change switches from managing Airbrake directly to using defra_ruby_alert, to

- make things better!
- be consistent with our other services (once we've done the same to them)
- and to hopefully finally ship RUBY-604!