Lock Byebug version (#321)
Dependabot tried to update byebug to 11.1.0 here: https://github.com/DEFRA/flood-risk-engine/pull/320

However we noticed that this version drops support for Ruby 2.3, which this engine is still on.

So for now, we lock the version to stop this update. However we can unlock it when we upgrade our version of Ruby.