Return correct error for valid postcode without results (#384)
https://eaflood.atlassian.net/browse/RUBY-1341

Previously if we searched for a postcode with no results, it would display an error message saying that the address finder was not working. This PR fixes the issue so we instead say that no results were found.

I had a look at the spec for this and it was wrong, as it was trying to raise a 400 error which is no longer the response we expect, and then testing for the wrong error message anyway.