Save blank instead of nil when populating address org (#383)
https://eaflood.atlassian.net/browse/RUBY-1341

We spotted a bug where trying to save certain addresses would fail. This is because they had no organisation.

The existing schema expects there to be a non-nil value for the organisation. A blank string is acceptable and that is what was previously being substituted in, so we'll do the same here to avoid any unexpected breakage down the line.