Clean up duplicate address after address is changed (#385)
https://eaflood.atlassian.net/browse/RUBY-1341

During testing we discovered an issue where if an address was added via postcode lookup, and then changed later to one entered manually, the previous address would still be assigned and often displayed instead of the amended, manually-entered version.

To fix this, I've added a callback to clean up any duplicate addresses after a new one is added. The intent is to refactor these forms soon anyway so this is a temporary fix.