Set up new registration workflow with AASM
https://eaflood.atlassian.net/browse/RUBY-1471

Since the state machine gem that FRAE has been using is no longer supported and is blocking further upgrade work, we have decided to switch to the AASM workflow that we use on waste carriers and waste exemptions.

This PR includes setting up a basic NewRegistration model and defining the workflow that it should use.