Create NewRegistration model
This subclasses a TransientRegistration model for consistency with the other engines, and to leave room for future refactoring of other journeys (eg editing or renewal).

Since the focus is currently on workflow, this PR does not define all the various attributes and relations of the new registration, only the token and workflow state.