Set up new registration workflow with AASM (#390)
https://eaflood.atlassian.net/browse/RUBY-1471

Since the state machine gem that FRAE has been using is no longer supported and is blocking further upgrade work, we have decided to switch to the AASM workflow that we use on waste carriers and waste exemptions.

This PR includes setting up a basic NewRegistration model and defining the workflow that it should use. This subclasses a TransientRegistration model for consistency with the other engines, and to leave room for future refactoring of other journeys (eg editing or renewal).

Since the focus is currently on workflow, this PR does not define all the various attributes and relations of the new registration, only the token and workflow state.