Set up controllers for updated new reg workflow
https://eaflood.atlassian.net/browse/RUBY-1472

This PR sets up new controllers and routes for the updated new registration workflow. These will follow the format we already use in WCR and WEX.

The end result should be an empty 'happy path' journey that a user can click through and see the pages in the expected journey. The actual content of the pages won't be merged in yet.