Set up base
These are primarily based on WCR, although I have made some minor simplifications around the setup checks since FRAE isn't dealing with the same complex permissions.

Each form has a controller and a form object. When the post request hits the controller, the form is used to handle any validations and data saving, and then the controller uses the state machine to redirect the user to the next correct page (or not, if validation errors occur).