Set up controllers for updated new reg workflow (#393)
https://eaflood.atlassian.net/browse/RUBY-1472

This PR sets up new controllers and routes for the updated new registration workflow. These will follow the format we already use in WCR and WEX.

The end result should be an empty 'happy path' journey that a user can click through and see the pages in the expected journey. The actual content of the pages won't be merged in yet.

The base classes are primarily based on WCR, although I have made some minor simplifications around the setup checks since FRAE isn't dealing with the same complex permissions.

Each form has a controller and a form object. When the post request hits the controller, the form is used to handle any validations and data saving, and then the controller uses the state machine to redirect the user to the next correct page (or not, if validation errors occur).