Turn NewRegistrations into Enrollments on submission (#417)
https://eaflood.atlassian.net/browse/RUBY-1474

Once a user submits their application to register, we need to turn the NewRegistration record that has been used for the forms into a proper Enrollment record.

This also means restructuring the data to match the Enrollment schema and its related objects. Once the submission is completed, the new Enrollment should be visible and manageable in the back office as before.

The old temporary NewRegistration record should then be deleted.