Validate company number with Companies House (#435)
https://eaflood.atlassian.net/browse/RUBY-1506

Previously we just validated the format of the Companies House number to see if it looked about right. Now we use our validation gem to send it off to the Companies House API to check if it actually belongs to an active company.

Also ripped out some old validators that are no longer used.