Add timestamps to transient records
https://eaflood.atlassian.net/browse/RUBY-1685

While working on a ticket to remove old transient_registrations, I realised there were no timestamps on the records. So this PR amends the existing migrations to include them.

This will mean wiping and redoing the database on each environment but as we haven't rolled these out to production yet, it seemed like a cleaner way to do it than adding the timestamps as additional migrations.